live/requirements.txt
websocket-client
redis
pandas
requests